requirements.txt
fastapi
uvicorn
playwright
google-generativeai
beautifulsoup4
requests
fake-useragent
python-dotenv
pillow
aiofiles
